+++ src/Commands/GoToLine.ts
@@ -1,10 +1,11 @@
 import * as vscode from "vscode"
+import { Commands } from "./index"
 
 export default class GoToLine implements vscode.Command {
 	title = ""
 	tooltip = ""
 	arguments: Array<any> = []
-	command = "jestExplorer.goToLine"
+	command = Commands.GoToLine
 
 	constructor(line: number) {
 		this.arguments = [line]