+++ src/Commands/RunJestTest.ts
@@ -1,12 +1,14 @@
 import * as vscode from "vscode"
+import { Commands } from "."
 
 export default class RunJestTest implements vscode.Command {
 	title = ""
 	tooltip = ""
 	arguments: Array<any> = []
-	command = "jestRunner.runJestTest"
+	command = Commands.RunJestTest
 
 	constructor(fileUrl: string) {
 		this.arguments = [fileUrl]
 	}
+
 }