+++ src/Models/TestStatus.ts
@@ -1,0 +1,8 @@
+enum TestStatus {
+	NotExecuted,
+	Failed,
+	Passed,
+	Skipped
+}
+
+export default TestStatus