+++ src/Utils/PathHelper.ts
@@ -3,25 +3,29 @@ import * as path from "path"
 import * as fs from "fs"
 
 export default class PathHelper {
+	private event = new vscode.EventEmitter<string>()
+
+	private lastPackageJson: string | undefined
 
 	constructor() {
 	}
 
-	public static getRelativeFilePathToWorkspaceRoot(fileUrl: string, projectRoot: string): string {
+	public getRelativeFilePathToWorkspaceRoot(fileUrl: string, projectRoot: string): string {
 		let subFolder = fileUrl.split(projectRoot)
-
 		const folder = subFolder[1].replace(/\\/g, "/")
-
-		// "./test/Components/Oldyojio.test.tsx"
 		return `./${folder}`
 	}
 
-	public static getLastPackageJsonForFile(fileUrl: string): string | null {
+	public getLastPackageJsonForFile(fileUrl: string): string | null {
+		if (this.lastPackageJson) {
+			if (fs.existsSync(this.lastPackageJson)) {
+				return this.lastPackageJson
+			}
+		}
+
 		const projectRoot = vscode.workspace.rootPath!
 		let currentRoot = projectRoot
 		let lastPackageJson: null | string = null
-
-		// magic
 
 		let subFolder = fileUrl.split(projectRoot)
 
@@ -38,9 +42,22 @@ export default class PathHelper {
 			}
 		})
 
-		// if (lastPackageJson && path.sep === "\\") {
-		// 	return lastPackageJson!.replace(/\\/g, "\\\\")
-		// }
+		if (!lastPackageJson) {
+			this.event.fire()
+		}
+
 		return lastPackageJson
+	}
+
+	public escapeWindowsString(str: string): string {
+		return this.escapeWindowsSpaces(this.escapeWindowsBackslashes(str))
+	}
+
+	public escapeWindowsSpaces(str: string): string {
+		return str.replace(/([^`])(\s)/g, (a, b, c) => `${b}\`${c}`)
+	}
+
+	public escapeWindowsBackslashes(str: string): string {
+		return str.replace(/\\/g, "\\\\")
 	}
 }